Add files via upload
Step 1: 

Declare a variable age using the var keyword and set it to the number 10.

Step 2: 

Add an "if" statement that checks if the value of the age variable is greater than or equal to the number 65. Inside the if block, type "console.log" and "You get your income from your pension".